pyproject.toml
[project]
name = "modern-data-bitcoin"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.11.11",
    "duckdb>=1.1.3",
    "numpy>=2.2.2",
    "pandas>=2.2.3",
    "plotly>=5.24.1",
    "polars>=1.20.0",
    "pyarrow>=19.0.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "shiny>=1.2.1",
    "shinywidgets>=0.4.2",
]
